pyproject.toml
[project]
name = "vigor"
version = "0.2.2"
description = "Vigor is a tool for generating and managing project templates."
requires-python = ">=3.8"
dependencies = [
    "click>=8.1.7",
    "filedate>=2.0",
    "GitPython>=3.1.14",
    "hurry.filesize>=0.9",
    "pillow>=11.0.0",
    "pypdf>=5.0.1",
]
authors = [
    {name = "Ryan Liu", email = "ryanliu6@pm.me"},
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
readme = "README.md"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.3.0",
    "mypy>=1.14.0",
    "types-click>=7.0",
    "types-PyYAML>=6.0.12",
    "types-Pygments>=2.15.0",
    "pre-commit>=3.5.0",
    "tomli>=2.0.1",
    "tomli-w>=1.0.0",
    "pyyaml>=6.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["vigor"]

[tool.ruff]
line-length = 88
target-version = "py38"

[tool.ruff.format]
quote-style = "double"
line-ending = "lf"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "F",  # pyflakes
    "B",  # flake8-bugbear
    "I",  # isort
    "UP", # pyupgrade
    "RUF", # Ruff-specific rules
]
ignore = [
    "D100",  # Missing docstring in public module
    "D104",  # Missing docstring in public package
    "D107",  # Missing docstring in __init__
    "D203",  # 1 blank line required before class docstring
    "D213",  # Multi-line docstring summary should start at the second line
    "D406",  # Section name should end with a newline
    "D407",  # Missing dashed underline after section
    "D413",  # Missing blank line after last section
    "COM812",  # Missing trailing comma
    "ISC001",  # Single line import statements
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
